requirements.txt
annotated-types==0.7.0
iniconfig==2.0.0
packaging==24.2
parsy==2.1
pluggy==1.5.0
pydantic==2.10.3
pydantic-core==2.27.1
pytest==8.3.4
python-dotenv==1.0.1
pyyaml==6.0.2
typing-extensions==4.12.2
